feat(workflows): Decompose InputHookService
- Resolved all review/CI blockers across six test files: corrected `uint` and `params` typing, and replaced blocking task waits with bounded async waits. Production behavior is unchanged; edits remain…
- Static checks passed, including `git diff --check`, five-state-machine structure, stale-pattern audit, and plan-file integrity. .NET build/tests were `not_run` because `dotnet` is unavailable.

- Files changed: 6
- Lines changed: +11 / -12
- New env vars: none
- Database migrations: none

+++ Tests/AutoRunStateMachineTests.cs
@@ -40,7 +40,7 @@ public sealed class AutoRunStateMachineTests
         machine.JoinBackgroundInputThread();
 
         var post = Assert.Single(transport.Posts);
-        Assert.Equal(NativeMethods.WM_KEYDOWN, post.Message);
+        Assert.Equal((uint)NativeMethods.WM_KEYDOWN, post.Message);
         Assert.Equal(0x57, post.VirtualKey);
     }
